Remove implicit queueing
There are 2 cases where commands are being queued: when using
`#combine_options` and when using `#method_missing` methods (like
`#resize`). This commit drops the `#method_missing` queueing.

The reason for this is that the code becomes much more complicated when
at the beginning of each "terminal" method we have to run the queue if
there are any queued commands. But when we write `image.resize("150x150")`,
we don't expect it to be queued, and it's not intuitive either. I can
imagine a user writing the following code, and being suprised it isn't
working:

  path = "path/to/image.jpg"
  image = MiniMagick::Image.new(path)
  image.resize("150x150")

  # User: "Done! The image is cropped."

We already have a great API for queueing commands -- `#combine_options`,
which I'm sure users have already adopted, and that no-one will be
suprised if the regulard commands aren't queueing anymore.

I will thoroughly update the README before the next release.

Fixes #230.