Add respond_to_missing? to complement method_missing
The `#respond_to_missing?` hook is the complement of `#method_missing`,
used by `#respond_to?`, `#method`, etc... to do the right thing. In this
case, since `#method_missing` is implemented in terms of an instance of
`MiniMagick::Tool::Mogrify`, we use an instance of that class to
determine if objects of this class can respond to those missing
messages.