Delete *.cache files generated with .mpc format
ImageMagick generates *.cache files when dealing with the .mpc format.
These files can be fairly large so it's important that they're deleted
when we're destroying the image.

From the ImageMagick documentation it seems that the .mpc format is the
only one for which the *.cache files are generated. So we want to issue
a `rm` command only with .mpc files, in order to not access the
filesystem unnecessarily. We also use the `-f` option so that if for
some reason the *.cache file doesn't exist there is no error.

Fixes #358