Parse out Ghostscript warnings in identfication
Image info sometimes parses incorrectly because Ghostscript may output
warnings about the validity of the object, despite successfully reading
its attributes. This causes parsing problems because the warning
messages are concatenated with the attributes, and trying to access
these attributes via MiniMagick results in type errors.

The fix is to identify if the warning messages appear, and if so, match
the attributes via strict regular expression.