Use `magick` instead of `magick convert` on IM 7
It appears that `magick convert` has been explicitly deprecated in
newest versions of IM 7, so we switch to using `magick` in that case. I
don't know how much `magick` is compatible with `magick convert`, but we
have to avoid the deprecation warning, so hopefully we can make the
switch.

The `convert` command seems to be the only one that was explicitly
deprecated, `magick identify` and others still appear to work.