Ignore warnings about convert command being deprecated on IM7
I don't think `magick` is 100% backwards compatible with `magick
convert`, so for now it's better to ignore the warning so that people
aren't spammed.

The next release will most likely be a major release, so we'll use it to
switch to `magick` on IM7. That way people will be more careful in
checking whether their processing commands still work the same.

Closes #560