Remove official support for GraphicsMagick
Supporting GraphicsMagick adds a lot of complexity to MiniMagick, due to
lack of support for some features. For example, I wanted to use native
ImageMagick time limit for command timeout, but GraphicsMagick doesn't
support it.

We remove support for it, but document how to use it via `#cli_prefix`
if needed.

This removes `#cli` and `#processor` configurations, as they're not
needed anymore. We also remove `#cli_path`, because it isn't documented
and seems like it never worked in recent versions of MiniMagick, because
CLI detection didn't take it into account.